big update (too lazy to explain)

+++ server/api/journal.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
-const {v4 : uuidv4} = require('uuid')
+const { v4: uuidv4 } = require('uuid')
 
 const express = require("express");
 const router = express.Router();
@@ -36,7 +36,7 @@ const documentBucket = getStorage().bucket();
 
 // CREATE JOURNAL ENTRIES
 
-router.post('/new-entry', upload.single('file'), async (req, res) => {
+router.post('/new-entry', upload.single('file'), authUser, authAccountant(ROLE.MANAGER, ROLE.BASIC), async (req, res) => {
 
   var {
     transactions,
@@ -78,29 +78,31 @@ router.post('/new-entry', upload.single('file'), async (req, res) => {
   const journalID = (counter.data().count + 1);
 
 
- try {
-  await journalRef.doc("" + journalID).set({
-    id: journalID,
-    transactions,
-    desc,
-    date,
-    userName,
-    status: "pending"
-  })
-    .then(async (res) => {
-      const files = fs.readdirSync('uploads/')
-
-      await documentBucket.upload(`uploads/${files[0]}`, {
-        destination: `journalDocuments/${journalID}/${files[0]}`, public: true, metadata: {
-          metadata: {
-            firebaseStorageDownloadTokens: uuidv4(),
-          }
+  try {
+    await journalRef.doc("" + journalID).set({
+      id: journalID,
+      transactions,
+      desc,
+      date,
+      userName,
+      status: "pending"
+    })
+      .then(async (res) => {
+        const files = fs.readdirSync('uploads/')
+        console.log(files.length === 0);
+        if (files.length !== 0) {
+          await documentBucket.upload(`uploads/${files[0]}`, {
+            destination: `journalDocuments/${journalID}/${files[0]}`, public: true, metadata: {
+              metadata: {
+                firebaseStorageDownloadTokens: uuidv4(),
+              }
+            }
+          })
+          fs.unlink(`uploads/${files[0]}`, (err) => {
+            console.log(err)
+          });
         }
-      })
-      fs.unlink(`uploads/${files[0]}`, (err) => {
-        console.log(err)
       });
-    });
     await eventLog.saveEventLogCreateJournal(req, journalID.toString());
     res.json('Successfully added journal');
   } catch (error) {
@@ -160,6 +162,7 @@ router.put('/entry/approve/:entryID', authUser, authRole(ROLE.MANAGER), async (r
   }
 
   var entry = fetchID.data();
+  var transactions = entry.transactions;
 
   if (entry.status !== "pending") {
     return res.status(400).json({ errors: "Journal not pending" });
@@ -170,26 +173,27 @@ router.put('/entry/approve/:entryID', authUser, authRole(ROLE.MANAGER), async (r
   var accountData;
   var accountRef;
 
-  for (var transaction of entry.transactions) {
+  for (var x = 0; x < transactions.length; x++) {
     try {
-      accountRef = accountsRef.doc(transaction.accountID);
+      console.log(transactions[x]);
+      accountRef = accountsRef.doc(transactions[x].accountID);
       accountDb = await accountRef.get();
       accountData = accountDb.data();
       var oldBalance = accountData.balance
+      console.log(accountData);
 
-      accountData.credit += transaction.creditAmount;
-      accountData.debit += transaction.debitAmount;
+      transactions[x].creditAfter = accountData.credit + transactions[x].creditAmount;
+      transactions[x].debitAfter = accountData.debit + transactions[x].debitAmount;
 
-      transaction.creditAfter = accountData.credit;
-      transaction.debitAfter = accountData.debit;
+      console.log(accountData);
 
       if (accountData.normalSide === "L" || accountData.normalSide === "l") {
-        accountData.balance += transaction.debitAmount;
-        accountData.balance -= transaction.creditAmount;
+        accountData.balance += transactions[x].debitAmount;
+        accountData.balance -= transactions[x].creditAmount;
       }
       else if (accountData.normalSide === "R" || accountData.normalSide === "r") {
-        accountData.balance += transaction.creditAmount;
-        accountData.balance -= transaction.debitAmount;
+        accountData.balance += transactions[x].creditAmount;
+        accountData.balance -= transactions[x].debitAmount;
       }
       const updateAccount = {
         name: accountData.name,
@@ -212,15 +216,15 @@ router.put('/entry/approve/:entryID', authUser, authRole(ROLE.MANAGER), async (r
         category: accountData.category,
         subcategory: accountData.subcategory,
         balance: accountData.balance,
-        credit: accountData.credit + transaction.creditAmount,
-        debit: accountData.debit + transaction.debitAmount,
+        credit: accountData.credit + transactions[x].creditAmount,
+        debit: accountData.debit + transactions[x].debitAmount,
         assignedUsers: accountData.assignedUsers,
         comment: accountData.comment,
         statement: accountData.statement
       };
 
       batch.update(accountRef, newAccount);
-      await eventLog.saveEventLogUpdate(req, res, transaction.accountID, updateAccount, newAccount);
+      await eventLog.saveEventLogUpdate(req, res, transactions[x].accountID, updateAccount, newAccount);
 
     }
     catch (e) {
